+++ 3rdpj/src/main/java/com/goott/pj3/board/review/controller/ReviewController.java
@@ -1,6 +1,5 @@
 package com.goott.pj3.board.review.controller;
 
-import com.goott.pj3.common.util.FileUploadUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
@@ -8,12 +7,9 @@ import com.goott.pj3.board.review.service.ReviewService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;
