--- 3rdpj/src/main/webapp/WEB-INF/views/plan/uploadResult.jsp
@@ -1,20 +1,0 @@
-<%--
-  Created by IntelliJ IDEA.
-  User: Ang
-  Date: 2023-04-10
-  Time: 오전 12:37
-  To change this template use File | Settings | File Templates.
---%>
-<%@ page contentType="text/html;charset=UTF-8" language="java" %>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>File Upload Result</title>
-</head>
-<body>
-<h1>File Upload Result</h1>
-<p>${message}</p>
-<p><a href="/uploadForm">Upload another file</a></p>
-</body>
-</html>
